pyproject.toml
[tool.poetry]
name = "tftk"
version = "0.1.11"
description = "Machine Learning Toolkit using TensorFlow"
authors = ["Naruhide KITADA <kitfactory@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
gym = "^0.17.1"
optuna = "^1.2.0"
pillow = "^7.0.0"
tensorflow_probability = "^0.9.0"
icrawler = "^0.6.2"
graphviz = "^0.14"
pydot-ng = "^2.0.0"
tensorboard_plugin_profile = "^2.2.0"
tensorflow_datasets = "^3.1.0"
tensorflow = "^2.2.0"
tensorflow_addons = "^0.10.0"
sphinx = "^3.0.3"
sphinx_ustack_theme = "^1.0.0"
sphinx_materialdesign_theme = "^0.1.11"
matplotlib = "^3.2.1"
opencv-python = "^4.2.0"
pymongo = "^3.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
